Show how to write matplotlib tests with image comparison (close #106)
- [ ] Eliminate inconsistencies between Python 3.6 and other versions
- [ ] Ensure that setting of Agg backend works, and that there are no
      crashes on Travis

Added nose to the conda environment, because matplotlib uses it in its
testing infrastructure.

Added pytest-mpl to the conda environment, because it facilitates
using matplotlib's testing infrastructure in pytest.

Note that the matplotlib backend must be set to Agg *before* impolting
`matplotlib.pyplot` in order to avoid crashes in headless
environments.

It seems that the labels on the y-axis are rendered slightly higher in
Python 3.6 than in Pythons 3.5 and 2.7. Consequently the example test
fails in Python 3.6.

**NOTE: agg backend setting still doesn't seem to work correctly** in
spite of having tried a couple of solutions.